requirements.txt
# OpenAI Agents SDK and dependencies
openai>=1.0.0
openai-agents-sdk>=0.1.0

# HTTP and API clients
requests>=2.31.0
aiohttp>=3.9.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Geospatial calculations
geopy>=2.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
typing-extensions>=4.8.0
